Move createAccessToken to obtain-access-token.js
