+++ test/setup-request/setup-request.test.js
@@ -8,20 +8,11 @@ const authorisationServerId = 'testAuthorisationServerId';
 describe('accessTokenAndResourcePath called with valid parameters', () => {
   const token = 'access-token';
   const resourceServerPath = 'http://resource-server.com/open-banking/v1.1';
-  const clientId = 'testClientId';
-  const clientSecret = 'testClientSecret';
-  const tokenPayload = {
-    scope: 'accounts payments',
-    grant_type: 'client_credentials',
-  };
-  const tokenResponse = { access_token: token };
-  const tokenStub = sinon.stub().returns(tokenResponse);
-  const getClientCredentialsStub = sinon.stub().returns({ clientId, clientSecret });
+  const tokenStub = sinon.stub().returns(token);
   const resourceServerPathStub = sinon.stub().returns(resourceServerPath);
   const accessTokenAndResourcePathProxy = proxyquire('../../app/setup-request/setup-request', {
-    '../authorise': { postToken: tokenStub },
+    '../authorise': { createAccessToken: tokenStub },
     '../authorisation-servers': {
-      getClientCredentials: getClientCredentialsStub,
       resourceServerPath: resourceServerPathStub,
     },
   }).accessTokenAndResourcePath;
@@ -29,10 +20,7 @@ describe('accessTokenAndResourcePath called with valid parameters', () => {
   it('returns accessToken and resourcePath', async () => {
     const { accessToken, resourcePath } =
       await accessTokenAndResourcePathProxy(authorisationServerId);
-    assert(tokenStub.calledWithExactly(
-      authorisationServerId,
-      clientId, clientSecret, tokenPayload,
-    ), 'postToken called correctly');
+    assert(tokenStub.calledWithExactly(authorisationServerId), 'postToken called correctly');
 
     assert.equal(resourcePath, resourceServerPath);
     assert.equal(accessToken, token);
